❖	Image Search Application (Kotlin)
It’s a simple app that uses Unsplash photo  api to get photos by their engine and show them on recycler view in a great view.

When you search on something, it gets all the photos about the searched work & the Photographer who took it or the designer who designed it.

When clicking the image it opens a detailed screen with all information about the photo.