❖	Pokegmon Application (Java)
❖	Used: RXJava – Retrofit – MVVM -  Dagger Hilt – View binding – LiveData - Room

It’s a very simple app that uses an api to get images about famous Pokémon and their names.

When swipe the image(View) to the right, it’s get saved in the cache as a favorite

You can see your favorite images in specific view, also when swiping to the right, it’s get removed from the favorites(cache).

Link: https://drive.google.com/file/d/18W4QI-JJa_DEZkM0lJvyuFcL67uPYpKZ/view?usp=sharing